+++ server/src/tests/renameGlobalVarScope.spec.ts
@@ -1,70 +1,14 @@
 import * as lpc from "./_namespaces/lpc.js";
-import * as path from "path";
+import { createTestLanguageService } from "./harness.js";
 
-function makeLanguageService(files: Map<string, string>) {
-    const options: lpc.CompilerOptions = {
-        driverType: lpc.LanguageVariant.FluffOS,
-        diagnostics: true,
-    };
-
-    const scriptFiles = [...files.keys()];
-    const scriptVersions = new Map<string, string>(scriptFiles.map((f) => [f, "1"]));
-
-    const currentDirectory = lpc.normalizePath(process.cwd());
-    const useCaseSensitiveFileNames = lpc.sys.useCaseSensitiveFileNames;
-    const getCanonicalFileName = lpc.createGetCanonicalFileName(useCaseSensitiveFileNames);
-    const normalizeHostFileName = (name: string | undefined) => (name ? lpc.normalizePath(name) : name);
-
-    const host: lpc.LanguageServiceHost = {
-        getCompilationSettings: () => options,
-        getCurrentDirectory: () => currentDirectory,
-        getDefaultLibFileName: (opts) =>
-            lpc.combinePaths(currentDirectory, lpc.getDefaultLibFolder(opts), lpc.getDefaultLibFileName(opts)),
-        getIncludeDirs: () => [],
-        getParseableFiles: () => new Set(scriptFiles.map((f) => lpc.toPath(f, currentDirectory, getCanonicalFileName))),
-        getScriptFileNames: () => scriptFiles,
-        getScriptSnapshot: (name) => {
-            const normalizedName = normalizeHostFileName(name);
-            if (!normalizedName) return undefined;
-            const text = files.get(normalizedName) ?? lpc.sys.readFile(normalizedName);
-            return text === undefined ? undefined : lpc.ScriptSnapshot.fromString(text);
-        },
-        getScriptVersion: (name) => {
-            const normalizedName = normalizeHostFileName(name);
-            return normalizedName ? scriptVersions.get(normalizedName) ?? "0" : "0";
-        },
-        isKnownTypesPackageName: () => false,
-        useCaseSensitiveFileNames: () => useCaseSensitiveFileNames,
-        fileExists: (name) => {
-            const normalizedName = normalizeHostFileName(name);
-            return !!normalizedName && (files.has(normalizedName) || lpc.sys.fileExists(normalizedName));
-        },
-        readFile: (name) => {
-            const normalizedName = normalizeHostFileName(name);
-            return normalizedName ? files.get(normalizedName) ?? lpc.sys.readFile(normalizedName) : undefined;
-        },
-        onAllFilesNeedReparse: () => undefined,
-        onReleaseOldSourceFile: () => undefined,
-        onReleaseParsedCommandLine: () => undefined,
-    };
-
-    const fileHandler = lpc.createLpcFileHandler({
-        fileExists: (name) => host.fileExists(name),
-        readFile: (name) => host.readFile(name),
-        getIncludeDirs: () => host.getIncludeDirs(),
-        getCompilerOptions: () => host.getCompilationSettings(),
-        getCurrentDirectory: () => host.getCurrentDirectory(),
-    });
-
-    const ls = lpc.createLanguageService(host, fileHandler);
-
+function makeLanguageService(files: Record<string, string>) {
+    const svc = createTestLanguageService(files);
     // force check every file so LPC inherits are resolved
-    const program = ls.getProgram()!;
-    for (const f of scriptFiles) {
-        program.getTypeChecker().getDiagnostics(program.getSourceFile(f));
+    const program = svc.ls.getProgram()!;
+    for (const rel of Object.keys(files)) {
+        program.getTypeChecker().getDiagnostics(program.getSourceFile(svc.abs(rel)));
     }
-
-    return ls;
+    return svc;
 }
 
 function renameCountsByFile(ls: lpc.LanguageService, fileName: string, pos: number) {
@@ -78,8 +22,6 @@ function renameCountsByFile(ls: lpc.LanguageService, fileName: string, pos: numb
 
 describe("LanguageService rename global var scope", () => {
     it("does not rename same-named globals in unrelated files", () => {
-        const fileA = lpc.normalizePath(path.join(process.cwd(), "a.c"));
-        const fileB = lpc.normalizePath(path.join(process.cwd(), "b.c"));
         const sourceA = `int foo;
 
 test() {
@@ -94,7 +36,9 @@ other() {
     return foo;
 }
 `;
-        const ls = makeLanguageService(new Map([[fileA, sourceA], [fileB, sourceB]]));
+        const { ls, abs } = makeLanguageService({ "a.c": sourceA, "b.c": sourceB });
+        const fileA = abs("a.c");
+        const fileB = abs("b.c");
 
         const declPos = sourceA.indexOf("foo");
         const byFile = renameCountsByFile(ls, fileA, declPos);
@@ -104,8 +48,6 @@ other() {
     });
 
     it("renames inherited global var references in child files", () => {
-        const parent = lpc.normalizePath(path.join(process.cwd(), "parent.c"));
-        const child = lpc.normalizePath(path.join(process.cwd(), "child.c"));
         const sourceParent = `int foo;
 
 test() {
@@ -120,7 +62,9 @@ child_fn() {
     return foo;
 }
 `;
-        const ls = makeLanguageService(new Map([[parent, sourceParent], [child, sourceChild]]));
+        const { ls, abs } = makeLanguageService({ "parent.c": sourceParent, "child.c": sourceChild });
+        const parent = abs("parent.c");
+        const child = abs("child.c");
 
         const declPos = sourceParent.indexOf("foo");
         const byFile = renameCountsByFile(ls, parent, declPos);